added:CRUD operations in Experience

+++ services/experience.service.ts
@@ -11,15 +11,27 @@ class ExperienceService {
         return this.expRepo.addExperience(userId, experienceData);
     }
 
-    async updateExperience(userId: string, experienceId: string, experienceData: Partial<IExperience>) {
+  async update(
+    userId: string,
+    experienceId: string,
+    data: Partial<IExperience>
+  ) {
+    if (!experienceId) throw new Error("experienceId required");
+    return this.expRepo.updateExperience(userId, experienceId, data);
+  }
 
-        return this.expRepo.updateExperience(userId, experienceId, experienceData);
-    }
+  async delete(userId: string, experienceId: string) {
+    if (!experienceId) throw new Error("experienceId required");
+    return this.expRepo.deleteExperience(userId, experienceId);
+  }
 
-    async deleteExperience(userId: string, experienceId: string) {
-
-        return this.expRepo.deleteExperience(userId, experienceId);
-    }
+  async getByExperienceId(userId: string, experienceId: string) {
+    if (!experienceId) throw new Error("experienceId required");
+    return this.expRepo.getExperienceById(userId, experienceId);
+  }
+  // async getAll(userId: string) {
+  //   return this.expRepo.getExperiences(userId);
+  // }
 }
 
 export default new ExperienceService();
